refactor(web): back link fixes for ip comments and lpa statement flows
